chore: restructure parser files
